pyproject.toml
[tool.poetry]
name = "raspai"
version = "0.1.0"
description = ""
authors = ["oyako-li <oyakoli3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.2.2"
numpy = "^1.24.2"
opencv-python = "^4.7.0.68"
pynput = "^1.7.6"
pyserial = "^3.5"
python-xlib = "^0.33"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
